requirements.txt
requests>=2.25.0
openai>=0.27.0
google-generativeai>=0.3.0
pyinstaller>=5.0.0
